separate out getting index and getting freq - a function can't return two things sequentially
This won't work:
function a(){
	return thing1;
	return thing2;
}

The first return statement finishes execution of the function and passes
flow-control back to the calling code, so the second return statement is unreachable.

If you want to return two things from one function, you can put them temporarily
into a data structure (e.g. an array, or an object).

e.g. as an array:

function a(){
   return [thing1, thing2];
}

or as an object

function a(){
	return {
		freq: noteFreqs[ix],
		index: ix
	}
}